* use Fulfillment class instead of client.<endpoint>.method * fix linter errors
