+++ lib/authentication.js
@@ -10,23 +10,21 @@ const testAuth = async (z, bundle) => {
   let endpoint = 'https://api.connect.cloud.im/public/v1';
   let apiKey = bundle.authData.api_key;
   let message = '';
-  if (apiKey.indexOf('@') > -1){
-    const substr = apiKey.split('@');
-    apiKey = substr[0];
-    endpoint = substr[1];
-    message = `[${substr[1]}]`;
+  if (apiKey.indexOf('@') > -1) {
+    [apiKey, endpoint] = apiKey.split('@');
+    message = `[${endpoint}]`;
   }
-
+  /* eslint-disable no-param-reassign */
   bundle.authData.endpoint = endpoint;
   bundle.authData.api_key = apiKey;
   const client = getConnectClient(z, bundle);
   const response = await client.accounts.list();
   const account = response[0];
-  return Promise.resolve({ 
-    account_info: `${account.name} (${account.id}) ${message}`,
+  return Promise.resolve({
+    account_info: `${account.name} (${account.id}) ${message}`.trim(),
     account_type: account.type,
-    endpoint: endpoint,
-    actual_api_key:  apiKey
+    endpoint,
+    actual_api_key: apiKey,
   });
 };
 
@@ -35,14 +33,14 @@ const authentication = {
   test: testAuth,
   connectionLabel: '{{account_info}}',
   sessionConfig: {
-    perform: testAuth
+    perform: testAuth,
   },
   fields: [
     {
       key: 'endpoint',
       type: 'string',
       required: false,
-      computed: true
+      computed: true,
     },
     {
       key: 'api_key',
@@ -53,7 +51,7 @@ const authentication = {
     {
       key: 'actual_api_key',
       required: false,
-      computed: true
+      computed: true,
     },
     {
       key: 'account_type',
