+++ lib/creates/approve_request.js
@@ -4,12 +4,14 @@
  * @copyright (c) 2019. Ingram Micro. All Rights Reserved.
  */
 
+const { Fulfillment } = require('@cloudblueconnect/connect-javascript-sdk');
+
 const { getConnectClient } = require('../connect/http');
 const sample = require('../samples/request');
 
 const approveRequest = async (z, bundle) => {
-  const client = getConnectClient(z, bundle);
-  const response = await client.requests.approveWithTemplate(
+  const ff = new Fulfillment(getConnectClient(z, bundle));
+  const response = await ff.approveRequestWithTemplate(
     bundle.inputData.id,
     bundle.inputData.template_id,
   );
