+++ lib/creates/fill_fulfillment_parameters.js
@@ -3,13 +3,13 @@
  *
  * @copyright (c) 2019. Ingram Micro. All Rights Reserved.
  */
-
+const { Fulfillment } = require('@cloudblueconnect/connect-javascript-sdk');
 const { getConnectClient } = require('../connect/http');
 const sample = require('../samples/request');
 
 const updateRequestParameters = async (z, bundle) => {
-  const client = getConnectClient(z, bundle);
-  const response = await client.requests.updateRequestParameters(
+  const ff = new Fulfillment(getConnectClient(z, bundle));
+  const response = await ff.updateRequestParameters(
     bundle.inputData.request_id,
     bundle.inputData.params,
     bundle.inputData.note,
