+++ lib/creates/inquire_request.js
@@ -4,16 +4,17 @@
  * @copyright (c) 2019. Ingram Micro. All Rights Reserved.
  */
 
+const { Fulfillment } = require('@cloudblueconnect/connect-javascript-sdk');
 const { getConnectClient } = require('../connect/http');
 const sample = require('../samples/inquire_request');
 
 const inquireRequest = async (z, bundle) => {
-  const client = getConnectClient(z, bundle);
-  const response = await client.requests.inquireRequest(
+  const ff = new Fulfillment(getConnectClient(z, bundle));
+  const response = await ff.inquireRequestWithTemplate(
     bundle.inputData.request_id,
-    bundle.inputData.reason,
     bundle.inputData.template_id,
     bundle.inputData.params,
+    bundle.inputData.note,
   );
   return Promise.resolve(response);
 };
@@ -28,7 +29,7 @@ module.exports = {
   operation: {
     inputFields: [
       { key: 'request_id', label: 'Request ID', required: true },
-      { key: 'reason', label: 'Reason', required: true },
+      { key: 'note', label: 'Note', required: true },
       { key: 'template_id', label: 'Template ID', required: true },
       {
         key: 'params',
