+++ lib/creates/note_request.js
@@ -4,12 +4,13 @@
  * @copyright (c) 2019. Ingram Micro. All Rights Reserved.
  */
 
+const { Fulfillment } = require('@cloudblueconnect/connect-javascript-sdk');
 const { getConnectClient } = require('../connect/http');
 const sample = require('../samples/note_request');
 
 const noteRequest = async (z, bundle) => {
-  const client = getConnectClient(z, bundle);
-  const response = await client.requests.updateRequest(
+  const ff = new Fulfillment(getConnectClient(z, bundle));
+  const response = await ff.updateRequest(
     bundle.inputData.id,
     {
       note: bundle.inputData.note,
