+++ lib/creates/reject_request.js
@@ -3,14 +3,14 @@
  *
  * @copyright (c) 2019. Ingram Micro. All Rights Reserved.
  */
-
+const { Fulfillment } = require('@cloudblueconnect/connect-javascript-sdk');
 const { getConnectClient } = require('../connect/http');
 const sample = require('../samples/reject_request');
 
 const rejectRequest = async (z, bundle) => {
-  const client = getConnectClient(z, bundle);
-  const response = await client.requests.rejectRequest(
-    bundle.inputData.requestId,
+  const ff = new Fulfillment(getConnectClient(z, bundle));
+  const response = await ff.failRequest(
+    bundle.inputData.request_id,
     bundle.inputData.reason,
   );
   return Promise.resolve(response);
@@ -21,11 +21,11 @@ module.exports = {
   noun: 'Reject a Request',
   display: {
     label: 'Reject a Request',
-    description: 'Change the status of a Request to "rejected"',
+    description: 'Change the status of a Request to "failed"',
   },
   operation: {
     inputFields: [
-      { key: 'requestId', label: 'Request ID', required: true },
+      { key: 'request_id', label: 'Request ID', required: true },
       { key: 'reason', label: 'Reason', required: true },
     ],
     perform: rejectRequest,
