+++ test/creates/inquire_request.js
@@ -5,7 +5,7 @@
  */
 
 const should = require('should');
-const { RequestService } = require('@cloudblueconnect/connect-javascript-sdk/lib/connect/api');
+const { Fulfillment } = require('@cloudblueconnect/connect-javascript-sdk');
 const sinon = require('sinon');
 const zapier = require('zapier-platform-core');
 const responses = require('../responses');
@@ -34,7 +34,7 @@ describe('Connect Fulfillment Zapier App - Inquire Request', () => {
     };
 
     // Mock the sdk function to return this response 
-    sandbox.stub(RequestService.prototype, 'inquireRequest').returns(responses.creates.inquire_request);
+    sandbox.stub(Fulfillment.prototype, 'inquireRequestWithTemplate').returns(responses.creates.inquire_request);
     // Call to zapier function to test
     appTester(App.creates.inquire_request.operation.perform, bundle)
       .then(results => {
