+++ test/creates/note_request.js
@@ -5,7 +5,7 @@
  */
 
 const should = require('should');
-const { RequestService } = require('@cloudblueconnect/connect-javascript-sdk/lib/connect/api');
+const { Fulfillment } = require('@cloudblueconnect/connect-javascript-sdk');
 const sinon = require('sinon');
 const zapier = require('zapier-platform-core');
 const responses = require('../responses');
@@ -32,7 +32,7 @@ describe('Connect Fulfillment Zapier App - Change Note Request', () => {
     };
 
     // Mock the sdk function to return this response 
-    sandbox.stub(RequestService.prototype, 'updateRequest').returns(responses.creates.note_request);
+    sandbox.stub(Fulfillment.prototype, 'updateRequest').returns(responses.creates.note_request);
     // Call to zapier function to test
     appTester(App.creates.note_request.operation.perform, bundle)
       .then(results => {
